feature: Lego.
Lego is the ability to compose Svalboard firmwares like
legos.  Ballpoint and pointball disappear.  Because they
are just a trackball firmware and a trackpoint firmware,
flashed on the correct sides.

If there is a blank, it can not be master.

Point/point is now a vaild config.

DPI Setting for PS/2 is disabled until we figure out
how not to have it break the world.